@@ -1,15 +1,16 @@
 {
   "name": "crypto-bot-father",
-  "version": "1.5.0",
+  "version": "1.6.0",
   "description": "A crypto bot father for Telegram. The one that rule them all!",
   "main": "src/index.js",
+  "type": "module",
   "directories": {
     "doc": "docs"
   },
   "scripts": {
     "start": "node .",
-    "lint": "eslint src --ext .js",
-    "fix": "eslint src --ext .js --fix"
+    "lint": "eslint src",
+    "fix": "eslint src --fix"
   },
   "repository": {
     "type": "git",
@@ -32,15 +33,15 @@
     "body-parser": "^1.20.2",
     "dotenv": "^16.4.5",
     "express": "^4.18.3",
+    "http-errors-enhanced": "^3.0.0",
     "luxon": "^3.4.4",
     "node-telegram-bot-api": "^0.66.0",
+    "pino": "^9.5.0",
     "table": "^6.8.1"
   },
   "devDependencies": {
-    "eslint": "^8.57.0",
-    "eslint-config-standard": "^17.1.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-n": "^16.6.2",
-    "eslint-plugin-promise": "^6.1.1"
+    "eslint": "^9.15.0",
+    "neostandard": "^0.11.8",
+    "pino-pretty": "^13.0.0"
   }
 }
