@@ -1,8 +1,8 @@
-const express = require('express')
-const router = express.Router()
+import { Router } from 'express'
+const router = Router()
 
 router.get('/', (req, res) => {
   res.json({ version: '1.0', name: 'Crypto Bot Father', devs: [{ name: 'Melroy van den Berg', email: 'melroy@melroy.org' }] })
 })
 
-module.exports = router
+export default router
